Move pages and content to the top of the repo
This is more common in Next.js so it makes sense to follow this pattern.
It also makes the content a little easier to discover.
