--- content/pages/about.md
@@ -1,6 +1,0 @@
----
-layout: page
-title: About me
----
-
-This is page about me
