+++ pages/confirm-subscription.js
