+++ pages/resources.js
